Fix replacement to support multiple components
When one file has multiple @Component() declarations, the MagicString gets modified multiple times.

The way the MagicString works, though, the indices are all based on the *original* offsets. So when the componentRegex searched into the modified string, it would come up with offsets outside of the range of the original string, which would throw an error.

The simple fix is to search the *source* string instead.